Added basic class to inherit for tests
