+++ tests/routes/assets/test_assets.py
@@ -2,31 +2,10 @@ import unittest
 import json
 import io
 from datetime import datetime
-import os
+from tests.base import BaseTest
 
-from common.configuration import create_app
-
-os.environ["DATABASETYPE"] = "TESTING"
-os.environ["APP_SETTINGS"] = "common.base.TestingConfig"
-os.environ["TESTING"] = "true"
-
-
-class TestAssets(unittest.TestCase):
-    def setUp(self):
-        self.app, self.database = create_app()
-        self.app.config["current_db"] = self.database.db
-        self.database.init_db()
-        self.client = self.app.test_client()
-        json_data = {"username": "test1", "password": "test"}
-        jwt_response = self.client.post("/auth/signup", json=json_data)
-        self.jwt = json.loads(jwt_response.get_data(as_text=True)).get("jwt")
-
-    def tearDown(self):
-        self.database.drop_all_tables()
-        self.app = None
-        self.client = None
-        self.jwt = None
-
+class TestAssets(BaseTest):
+    
     def test_add_asset(self):
         current_date = datetime.now()
         acquired_date = current_date.strftime("%Y-%m-%d")
@@ -38,7 +17,8 @@ class TestAssets(unittest.TestCase):
             "jwt": self.jwt,
             "file": (io.BytesIO(b"this is a test"), "test.pdf"),
         }
-
+        
+        
         response = self.client.post(
             "/assets/asset_add", data=asset_data, content_type="multipart/form-data"
         )
